+++ Clipster/Clipster/Data/Protocol/Network/AuthService.swift
@@ -1,0 +1,3 @@
+protocol AuthService {
+    func loginWithApple(token: String) async
+}
